--- src/error-handler.js
@@ -1,7 +1,0 @@
-import { errorMessage } from './consts.js';
-
-export const errorHandler = (incomingError, targetError) => {
-    if (incomingError.code = targetError) {
-        throw Error(errorMessage);
-    }
-}
